+++ tests/testthat/test-bayes_mode.R
@@ -5,7 +5,6 @@ test_that("bayes_mode works with external MCMC output", {
   nu = c(5,5)
   p = c(0.8,0.2)
   params = c(eta = p, mu = mu, sigma = sigma, nu = nu)
-  pars_names = c("eta", "mu", "sigma", "nu")
   dist_type = "continuous"
   
   data = c(sn::rst(p[1]*1000, mu[1], sigma[1], nu = nu[1]),
@@ -18,8 +17,12 @@ test_that("bayes_mode works with external MCMC output", {
     sn::dst(x, pars["mu"], pars["sigma"], pars["xi"], pars["nu"])
   }
   
-  bayesmix = new_BayesMixture(fit, data, K = 2, burnin = 1, pars_names = pars_names,
-                                pdf_func = pdf_func, dist_type = dist_type)
+  bayesmix = new_BayesMixture(fit,
+                              data,
+                              K = 2,
+                              burnin = 1,
+                              pdf_func = pdf_func,
+                              dist_type = dist_type)
   
   bayesmode = bayes_mode(bayesmix)