+++ tests/testthat/test-discrete_MF.R
@@ -4,13 +4,14 @@ test_that("discrete_MF function returns expected results with dist = shifted_poi
   kappa = c(10,0)
   p = c(0.5,0.5)
   params = c(eta = p, lambda = lambda, kappa = kappa)
-  pars_names = c("eta", "lambda", "kappa")
   dist = "shifted_poisson"
   
   data = c(rpois(p[1]*1e3, lambda[1]) + kappa[1],
            rpois(p[2]*1e3, lambda[2]) + kappa[2])
   
-  modes = discrete_MF(params, data = data, pars_names = pars_names, dist = dist)
+  mix = new_Mixture(params, data = data, dist = dist)
+  modes = discrete_MF(mix)
+  
   expect_equal(modes[1] == 0,  TRUE)
   expect_equal(modes[2] == 1,  TRUE)
   expect_equal(modes[3] == 10,  TRUE)
@@ -21,13 +22,13 @@ test_that("discrete_MF function returns expected results with dist = poisson", {
   lambda = c(0.1,10)
   p = c(0.5,0.5)
   params = c(eta = p, lambda = lambda)
-  pars_names = c("eta", "lambda")
   dist = "poisson"
   
   data = c(rpois(p[1]*1e3, lambda[1]),
            rpois(p[2]*1e3, lambda[2]))
   
-  modes = discrete_MF(params, data = data, pars_names = pars_names, dist = dist)
+  mix = new_Mixture(params, data = data, dist = dist)
+  modes = discrete_MF(mix)
   expect_equal(modes[1] == 0,  TRUE)
   expect_equal(modes[2] == 9,  TRUE)
 })
@@ -38,7 +39,6 @@ test_that("discrete_MF function returns expected results with arbitrary function
   size = c(20,0.5)
   p = c(0.5,0.5)
   params = c(eta = p, mu = mu, size = size)
-  pars_names = c("eta", "mu", "size")
 
   data = c(rnbinom(p[1]*1e3, mu = mu[1], size = size[1]),
            rnbinom(p[2]*1e3, mu = mu[2], size = size[2]))
@@ -47,7 +47,8 @@ test_that("discrete_MF function returns expected results with arbitrary function
     dnbinom(x, mu = pars["mu"], size = pars["size"])
   }
   
-  modes = discrete_MF(params, data = data, pars_names = pars_names, pmf_func = pmf_func)
+  mix = new_Mixture(params, data = data, pdf_func = pmf_func)
+  modes = discrete_MF(mix)
   
   expect_equal(modes[1] == 0,  TRUE)
   expect_equal(modes[2] == 18,  TRUE)